fix: camera preview to custom view
